requirements.txt
apache-airflow-providers-cncf-kubernetes>=9,<12
google-cloud-storage==2.17.0
wandb==0.17.9
